新功能：增加实验室新增或者修改时可修改或新增开放时间

+++ src/main/java/com/cuit9622/olms/mapper/LabMapper.java
@@ -31,7 +31,7 @@ public interface LabMapper extends BaseMapper<Lab> {
      * @param labId
      * @return
      */
-    List<Long> getWeekday(@Param("labId") Long labId);
+    List<Integer> getWeekday(@Param("labId") Long labId);
 }
